Refactor: Migrate user metadata to shared package and update components
- Moved user metadata to shared package for better reusability.
- Updated components in the showcase app to utilize the new userMeta from shared.
- Refactored footer, header, and various portfolio components to use shared components and metadata.
- Introduced new components: ApplicationLogo, GlowCard, LoadingSpinner, OtherLoginOption, ThemeProviders, and ThemeSwitch.
- Removed unused dependencies and updated package.json files accordingly.
- Enhanced theme switching functionality with next-themes integration.
