+++ apps/saas/app/layout.tsx
@@ -1,3 +1,4 @@
+import PrelineScript from '@/components/preline-script';
 import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
@@ -26,6 +27,7 @@ export default function RootLayout({
         <html lang="en">
             <body className={`${geistSans.variable} ${geistMono.variable} antialiased`}>
                 {children}
+                <PrelineScript />
             </body>
         </html>
     );
