+++ apps/saas/global.d.ts
@@ -1,0 +1,9 @@
+import type { IStaticMethods } from 'preline/dist';
+
+declare global {
+    interface Window {
+        HSStaticMethods: IStaticMethods;
+    }
+}
+
+export {};
