+++ apps/saas/package.json
@@ -10,6 +10,10 @@
         "lint": "next lint"
     },
     "dependencies": {
+        "@tailwindcss/forms": "^0.5.10",
+        "bcryptjs": "^3.0.2",
+        "jose": "^6.0.10",
+        "preline": "^3.0.1",
         "react": "^19.0.0",
         "react-dom": "^19.0.0",
         "next": "15.3.2"
