+++ apps/showcase/app/api/auth/login/route.ts
@@ -1,4 +1,4 @@
-import { user } from '@/app/metadata';
+import { userMeta } from '@repo/shared';
 import { NextResponse } from 'next/server';
 
 export async function GET() {
@@ -7,7 +7,7 @@ export async function GET() {
     // response json with token
     return NextResponse.json({
         token: token,
-        user,
+        user: userMeta.user,
         message: 'Login successful'
     });
 }
