+++ apps/showcase/components/blog/header.tsx
@@ -1,9 +1,15 @@
-import siteMetadata, { headerNavLinks } from '@/app/metadata';
+import { ApplicationLogo, userMeta as siteMetadata, ThemeSwitch } from '@repo/shared';
 import Link from 'next/link';
-import ApplicationLogo from '../shared/application-logo';
-import ThemeSwitch from '../theme-switch';
 import MobileNav from './mobile-nav';
 import SearchButton from './search-button';
+
+export const headerNavLinks = [
+    { href: '/', title: 'Home' },
+    { href: '/blog', title: 'Blog' },
+    { href: '/tags', title: 'Tags' },
+    { href: '/projects', title: 'Projects' },
+    { href: '/about', title: 'About' }
+];
 
 const Header = () => {
     let headerClass = 'flex items-center w-full bg-white dark:bg-gray-950 justify-between py-10';
