+++ apps/showcase/package.json
@@ -23,11 +23,8 @@
         "@emailjs/browser": "^4.3.3",
         "@headlessui/react": "^2.2.2",
         "@tailwindcss/forms": "^0.5.10",
-        "bcryptjs": "^3.0.2",
         "body-scroll-lock": "4.0.0-beta.0",
-        "jose": "^6.0.10",
         "next": "^15",
-        "next-themes": "^0.4.6",
         "nextjs-toploader": "^1.6.12",
         "preline": "^3.0.1",
         "react": "^19",
