+++ packages/shared/package.json
@@ -23,6 +23,8 @@
         "typescript": "5.8.2"
     },
     "dependencies": {
+        "@headlessui/react": "^2.2.2",
+        "next-themes": "^0.4.6",
         "react": "^19.1.0",
         "react-dom": "^19.1.0"
     }
