+++ packages/shared/src/components/application-logo.tsx
