+++ packages/shared/src/components/theme-providers.tsx
@@ -1,11 +1,10 @@
 'use client';
 
-import siteMetadata from '@/app/metadata';
 import { ThemeProvider } from 'next-themes';
 
 export function ThemeProviders({ children }: { children: React.ReactNode }) {
     return (
-        <ThemeProvider attribute="class" defaultTheme={siteMetadata.theme} enableSystem>
+        <ThemeProvider attribute="class" defaultTheme="light" enableSystem>
             {children}
         </ThemeProvider>
     );
