+++ packages/shared/src/index.tsx
@@ -1,4 +1,11 @@
 export { Button } from './button';
 export { Card } from './card';
 export { Code } from './code';
+export { default as ApplicationLogo } from './components/application-logo';
+export { default as GlowCard } from './components/glow-card';
+export { default as LoadingSpinner } from './components/loading-spinner';
+export { default as OtherLoginOption } from './components/other-login-option';
+export { ThemeProviders } from './components/theme-providers';
+export { default as ThemeSwitch } from './components/theme-switch';
+export { default as userMeta } from './metadata';
